requirements.txt
annotated-types==0.7.0
blinker==1.9.0
boto3==1.40.52
botocore==1.40.74
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.25.2
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gunicorn==22.0.0
grpcio==1.76.0
grpcio-status==1.71.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
packaging==24.2
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
PyPDF2==3.0.1
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
Werkzeug==3.1.3
